Qt/Debugger: Minor fixes

+++ Source/Core/DolphinQt2/Debugger/RegisterWidget.h
@@ -40,7 +40,7 @@ private:
   void OnItemChanged(QTableWidgetItem* item);
 
   void AddRegister(int row, int column, RegisterType type, std::string register_name,
-                   std::function<u32()> get_reg, std::function<void(u32)> set_reg);
+                   std::function<u64()> get_reg, std::function<void(u64)> set_reg);
 
   QTableWidget* m_table;
   bool m_updating = false;
